[FIX #371] Fixed bug when the Surface detection module by Lari crashed Regarding the module "Surface detection by Lari" crashes appeared sometimes. The reason was that the object containing the status bar thing (point classifier class) was deleted before the progress was set to be finished.
Remaining bug:
The following bug could not be fixed short-time because it does not appear if
the module is compiled using the debug flag.
Shortly said sometimes the search instance even can not find a point using
coordinates of itself. Weird. I put setting the ID in the end of
WLariBruteforceClustering::addExtentCluster manually at least not to get stuck.
But still some groups result a bad result. No errors appear when compiling (cmake)
using the debug flag. Has to be fixed sometime later.